feat: 集成 Spring Authorization Server 并支持 OAuth2 Password Grant
- 集成 Spring Authorization Server 1.3.4
- 新增 OAuth2 Password Grant 认证支持
- 本地账号登录统一走 /oauth2/token 端点
- 新增 OAuth2 核心配置类:
  - AuthorizationServerConfig
  - ResourceOwnerPasswordAuthenticationProvider
  - ResourceOwnerPasswordAuthenticationFilter
  - ClientPrincipalAuthenticationToken
- 添加 Knife4j OAuth2 认证配置 (OpenApiConfig)
- AuthController/GraphController 添加 OpenAPI 注解
- 接口请求/响应改用强类型 DTO/VO (替代 Map)
- Spring Boot 降级到 3.3.10 (Knife4j 兼容性)
- 前端 Login.vue 适配 OAuth2 token 端点

+++ veingraph-server/src/main/java/com/veingraph/rag/EsHybridSearchService.java
@@ -123,8 +123,8 @@ public class EsHybridSearchService {
                 s.knn(k -> {
                     k.field("vector")
                      .queryVector(toList(queryVector))
-                     .k(topK)
-                     .numCandidates(topK * 2)
+                     .k((long) topK)
+                     .numCandidates((long) topK * 2)
                      .boost(0.7f);
                     if (knnFilter != null) {
                         k.filter(knnFilter);
